pyproject.toml
[project]
name = "agentic-ai-w-fastmcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=2.14.4",
    "httpx>=0.28.1",
    "logfire>=4.20.0",
    "logging>=0.4.9.6",
    "pydantic>=2.12.5",
    "pydantic-ai>=1.47.0",
    "pymongo>=4.16.0",
    "python-dotenv>=1.2.1",
]
